feat: added quick-scope for highlightin a letter for easier find
We added the moded keys in the after folder,
we need to check if there is a better way to handle this.
We done this because we want to highlight only when we press f,t,F,T
The downside of this is because now when we want to go to second letter
the highlight wont work because count is first. (second letter is highlighted
with different color)